Initialize provider mapping (#9)
Add tables and functions that recommend which Pulumi provider to use for a given Terraform provider.

Relates pulumi/pulumi-service#35115

The initial pass matches versions 'up-to-latest-major'. And there is no pipeline to keep this up to date. In review we found that we need more accurate version pair matching and keep maintaining that. So this does not quite fix 35115.