Model migration state explicitly (#10)
For larger programs explicit migration mapping is very important as it keeps an explicit track of which Pulumi resource is representing a given Terraform resource, and also if some resources are skipped (this can be enhanced by a skip reason). Having this explicit also enables a diff command that can compare a Pulumi project to a Terraform project and correlates resources to check if they are "fully migrated". The peculiarities of the diff command give rise to explicit skips here for resources that need update or replace.