Implied type (#14)
ImpliedTtype produces a cty.Type from a shim.SchemaMap. This is necessary to correctly deserialize the terraform state into a cty.Value, as the state format loses some type information.

htype2ctype is directly copied from the bridge as I don't think we want to expose that publicly - we'll eventually get rid of it once we merge this into the bridge.

This adds some bridge-y code here, as I don't think it's good enough to go in the bridge. It works for the happy case which is why I am OK with committing it but we need to revisit it later. In particular, I doubt the DynamicType code is fully correct but we need a PF example to test it.

Note this is based on the following bridge PRs:
pulumi/pulumi-terraform-bridge#3270
pulumi/pulumi-terraform-bridge#3271
pulumi/pulumi-terraform-bridge#3274

part of pulumi/pulumi-service#34617