Work around failure to read TF-provisioned states (#52)
Work around failure to read TF-provisioned states.

This updates the loader to auto-detect when tofu show -json fails to resolve providers from Terraform registry and attempt to compensate for this by using a temp workspace to move the providers to OpenTofu providers and emit the translated OpenTofu state. When the workaround is applying the tool is emitting information on stderr explaining what it is doing.

Care is taken to not disrupt the project for users that prefer terraform, for example the lock file is saved and restored to avoid being edited by the workaround.

Fixes pulumi/pulumi-service#34864