Support reading remote TF state
The translate command used to assume that a `terraform.tfstate` file is provided as an input. Instead, the updated
version takes a path to Terraform sources and extracts the state from the current workspace regardless of which storage
backend is used, and auto-applies workarounds to parse Terraform provided states with Tofu.

The command is also cleaned up a little bit to have this new signature:

  pulumi-terraform-migrate stack \
    --from path/to/terraform-sources \
    --to path/to/pulumi-project \
    --out /tmp/pulumi-state.json \
    --plugins /tmp/required-plugins.json